hook up typescript frontend to backend
