+++ EmbeddedTerminal/WebView/default.html
@@ -8,10 +8,9 @@
     <title>Terminal Window</title>
     <link rel="stylesheet" href="../node_modules/xterm/dist/xterm.css" />
     <link rel="stylesheet" href="default.css" data-plugin-theme="true" />
-    <!-- <script src="../node_modules/xterm/dist/xterm.js" type="text/javascript"></script>
-    <script src="../node_modules/xterm/dist/addons/fit/fit.js" type="text/javascript"></script> -->
     <script src="../node_modules/requirejs/require.js"></script>
     <script>
+        function triggerEvent(event, data) { }
         requirejs.config({
             paths: {
                 'xterm': '../node_modules/xterm/dist/xterm',
@@ -28,7 +27,5 @@
     <div id="debug">
         debug:
     </div>
-    <!--<script src="linkMatcher.js" type="text/javascript"></script>
-    <script src="default.js" type="text/javascript"></script>-->
 </body>
 </html>
