+++ EmbeddedTerminal/WebView/ts/main.ts
@@ -1,18 +1,24 @@
 ﻿import { TermView } from './TermView';
 import { VisualStudio } from './VsEventManager';
 
-window.triggerEvent = (event, data) => {
+triggerEvent = function (event: string, data: any) {
     if (VisualStudio.isEventType(event)) {
         VisualStudio.Events.triggerEvent(event, data);
     }
 };
 
-console.log(document.readyState);
 if (document.readyState !== 'loading') {
-    let termView = new TermView('Fira Code', 14, 'C:\\');
+    let termView = new TermView(
+        JSON.parse(window.external.GetTheme()),
+        window.external.GetFontFamily(),
+        window.external.GetFontSize(),
+        window.external.GetSolutionDir());
 } else {
     document.addEventListener("DOMContentLoaded", function (event) {
-        console.log('event load fired');
-        let termView = new TermView('Fira Code', 14, 'C:\\');
+        let termView = new TermView(
+            JSON.parse(window.external.GetTheme()),
+            window.external.GetFontFamily(),
+            window.external.GetFontSize(),
+            window.external.GetSolutionDir());
     });
 }
