+++ EmbeddedTerminal/package.json
@@ -6,7 +6,6 @@
   "dependencies": {
     "node-pty": "^0.7.2",
     "requirejs": "^2.3.5",
-    "systemjs": "^0.21.0",
     "vscode-jsonrpc": "^3.5.0",
     "xterm": "=3.1.0"
   }
